requirements.txt
httpx==0.21.1
Pillow==8.4.0
requests==2.26.0
typer==0.4.0
